pyproject.toml
[tool.poetry]
name = "psychonaut"
version = "0.0.11"
description = "Python async client and TUI for bsky"
authors = ["generativist <jbn@abreka.com>"]
readme = "README.md"
homepage = "https://github.com/jbn/psychonaut"
repository = "https://github.com/jbn/psychonaut"
documentation = "https://github.com/jbn/psychonaut"
license = "MIT"

[tool.poetry.scripts]
psychonaut = "psychonaut.cli.main:cli"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.4"
pydantic = "^1.10.7"
click = "^8.1.3"
python-dotenv = "^1.0.0"
multiformats = "^0.2.1"
python-dateutil = "^2.8.2"
aiodns = "^3.0.0"
more-itertools = "^9.1.0"
websockets = "^11.0.2"
cbor2 = "^5.4.6"
dag-cbor = "^0.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
jmespath = "^1.0.1"
pytest-asyncio = "^0.21.0"
jinja2 = "^3.1.2"
coverage = {extras = ["toml"], version = "^7.2.3"}
ipykernel = "^6.22.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--capture=no"

[tool.coverage.run]
omit = [".*", "bin/*", "**/*_test.py"]

[tool.coverage.report]
fail_under = 1
